refactor: Slider 컴포넌트 분리
